+++ src/java/by/epam/ultraair/persistence/domain_objects/Client.java
@@ -1,4 +1,4 @@
-package by.epam.ultraair.service.domain_objects;
+package by.epam.ultraair.persistence.domain_objects;
 
 public class Client {
     private String name;
