+++ src/java/by/epam/ultraair/persistence/mappers/ClientMapper.java
@@ -1,7 +1,7 @@
-package by.epam.ultraair.service.mappers;
+package by.epam.ultraair.persistence.mappers;
 
 import by.epam.ultraair.dao.entities.ClientEntity;
-import by.epam.ultraair.service.domain_objects.Client;
+import by.epam.ultraair.persistence.domain_objects.Client;
 
 public class ClientMapper {
