+++ src/java/by/epam/ultraair/persistence/services/FlightService.java
@@ -1,10 +1,10 @@
-package by.epam.ultraair.service.services;
+package by.epam.ultraair.persistence.services;
 
 import by.epam.ultraair.presentation.dto.FlightDTO;
-import by.epam.ultraair.service.Constants;
-import by.epam.ultraair.service.domain_objects.Flight;
-import by.epam.ultraair.service.mappers.FlightIndexIdMapper;
-import by.epam.ultraair.service.mappers.FlightMapper;
+import by.epam.ultraair.persistence.Constants;
+import by.epam.ultraair.persistence.domain_objects.Flight;
+import by.epam.ultraair.persistence.mappers.FlightIndexIdMapper;
+import by.epam.ultraair.persistence.mappers.FlightMapper;
 import by.epam.ultraair.dao.entities.FlightEntity;
 import by.epam.ultraair.dao.repositories.FlightRepository;
 import org.json.simple.parser.ParseException;
